automl_backend_claude_v1/ml_pipeline/serializers.py
from rest_framework import serializers
from .models import Dataset, Model

class DatasetSerializer(serializers.ModelSerializer):
    class Meta:
        model = Dataset
        fields = '__all__'

class ModelSerializer(serializers.ModelSerializer):
    class Meta:
        model = Model
        fields = '__all__'