requirements_general.txt
Django>=3.2,<4.0
djangorestframework
psycopg2-binary
psutil==5.*
setuptools==2.*
typing-extensions==3.*
scipy==1.*
numpy==1.*
pandas==2.*
scikit-learn==1.*
category-encoders==2.*
xgboost==2.*
lightgbm==4.*
shap==0.44.*
django-cors-headers